@@ -1,4 +1,4 @@
-name: Almiya
+name: almiya
 description: A new Flutter project.
 
 # The following line prevents the package from being accidentally published to
@@ -18,13 +18,14 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
   font_awesome_flutter: ^8.8.1
+  provider: ^6.0.0
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
@@ -85,8 +86,11 @@ flutter:
   #         weight: 700
   #
   fonts:
-    - family: Almiya
+    - family: Almaya
       fonts:
-        - asset: assets/fonts/Almiya.ttf
+        - asset: assets/fonts/Almaya.ttf
+    - family: Assistant
+      fonts:
+        - asset: assets/fonts/Assistant-Regular.ttf
   # For details regarding fonts from package dependencies,
   # see https://flutter.dev/custom-fonts/#from-packages
