@@ -1,4 +1,4 @@
-/// Flutter icons Almiya
+/// Flutter icons Almaya
 /// Copyright (C) 2021 by original authors @ fluttericon.com, fontello.com
 /// This font was generated by FlutterIcon.com, which is derived from Fontello.
 ///
@@ -7,19 +7,19 @@
 ///
 /// flutter:
 ///   fonts:
-///    - family:  Almiya
+///    - family:  Almaya
 ///      fonts:
-///       - asset: fonts/Almiya.ttf
+///       - asset: fonts/Almaya.ttf
 ///
 /// 
 ///
 import 'package:flutter/widgets.dart';
 
-class Almiya {
-  Almiya._();
+class Almaya {
+  Almaya._();
 
-  static const _kFontFam = 'Almiya';
-  static const String _kFontPkg = null;
+  static const _kFontFam = 'Almaya';
+  static const String? _kFontPkg = null;
 
   static const IconData face = IconData(0xe800, fontFamily: _kFontFam, fontPackage: _kFontPkg);
   static const IconData asterisk = IconData(0xe801, fontFamily: _kFontFam, fontPackage: _kFontPkg);
