@@ -1,15 +1,16 @@
-import 'package:Almiya/pages/connect_to_places_page.dart';
-import 'package:Almiya/pages/information_main_page.dart';
-import 'package:Almiya/pages/messages_page.dart';
-import 'package:Almiya/pages/sign_up_complition_page.dart';
-import 'package:Almiya/pages/sign_up_page.dart';
-import 'package:Almiya/pages/welcome_page.dart';
-import 'package:Almiya/pages/login.dart';
-import 'package:Almiya/pages/what_you_need_page.dart';
+import 'package:almiya/pages/connect_to_places_page.dart';
+import 'package:almiya/pages/information_main_page.dart';
+import 'package:almiya/pages/messages_page.dart';
+import 'package:almiya/pages/sign_up_complition_page.dart';
+import 'package:almiya/pages/sign_up_page.dart';
+import 'package:almiya/pages/welcome_page.dart';
+import 'package:almiya/pages/login.dart';
+import 'package:almiya/pages/what_you_need_page.dart';
 import 'package:flutter/material.dart';
 
 import 'home_page.dart';
 import 'new_user_page.dart';
+import '../consts.dart';
 
 void main() {
   runApp(new MyApp());
@@ -20,6 +21,17 @@ class MyApp extends StatelessWidget {
   Widget build(BuildContext context) {
     return MaterialApp(
       title: 'Named Routes Demo',
+      theme: ThemeData(
+        elevatedButtonTheme: ElevatedButtonThemeData(
+          style: ElevatedButton.styleFrom(
+            primary: themeColor,
+            shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(15.0),
+            ),
+            onPrimary: buttonTextColor,
+          ),
+        ),
+      ),
       initialRoute: '/',
       routes: {
         '/': (context) => WelcomePage(),
